# public/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ApiResponse": {
				"properties": {
					"status": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {},
					"code": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"status",
					"message",
					"code"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "JWT Authorization header using the Bearer scheme."
			}
		}
	},
	"info": {
		"title": "boiler_api_5300",
		"version": "1.0.0",
		"description": "Boilerplate apis",
		"license": {
			"name": "license"
		},
		"contact": {
			"name": "STS"
		}
	},
	"paths": {
...
		"/libran/inventory/soft_delete_book_condition": {
			"delete": {
				"operationId": "DeleteBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "API: Soft Delete Book Condition\r\nMethod: DELETE\r\nURL: /soft_delete_book_condition\r\n \"bookId\": \"BOOK_ID\"",
				"tags": [
					"Library Book management"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"_id": {
										"type": "string"
									}
								},
								"required": [
									"_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/fine/add_fine": {
			"post": {
				"operationId": "AddFine",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"fine Book"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"isFineCancel": {
										"type": "number",
										"format": "double"
									},
									"paidDate": {
										"type": "string"
									},
									"isPaid": {
										"type": "number",
										"format": "double"
									},
									"returnDate": {
										"type": "string"
									},
									"book_id": {
										"type": "string"
									},
									"borrowId": {
										"type": "string"
									},
									"member_id": {
										"type": "string"
									}
								},
								"required": [
									"isFineCancel",
									"paidDate",
									"isPaid",
									"returnDate",
									"book_id",
									"borrowId",
									"member_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/fine/update_fine": {
			"put": {
				"operationId": "UpdateFine",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"fine Book"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"paidDate": {
										"type": "string"
									},
									"isFineCancel": {
										"type": "number",
										"format": "double"
									},
									"isPaid": {
										"type": "number",
										"format": "double"
									},
									"returnDate": {
										"type": "string"
									},
									"borrowId": {
										"type": "string"
									}
								},
								"required": [
									"paidDate",
									"isFineCancel",
									"isPaid",
									"returnDate",
									"borrowId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/fine/get_fine": {
			"get": {
				"operationId": "GetFine",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"fine Book"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/fine/delete_fine": {
			"delete": {
				"operationId": "DeleteFine",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"fine Book"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"_id": {
										"type": "string"
									}
								},
								"required": [
									"_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/libran/addBook/add_book": {
			"post": {
				"operationId": "AddBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "API: Add Book\r\nMethod: POST\r\nURL: /add_book\r\n\r\nBody:\r\n{\r\n  \"title\": \"Book Name\",\r\n  \"author\": \"Author Name\",\r\n  \"isbn\": \"1234567890\",\r\n  \"genre\": \"Fiction\",        \r\n  \"totalCopies\": 10,\r\n  \"shelfLocation\": \"A1\"      \r\n}",
				"tags": [
					"Librarian Add Book"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"shelfLocation": {
										"type": "string"
									},
									"totalCopies": {
										"type": "number",
										"format": "double"
									},
									"genre": {
										"type": "string"
									},
									"isbn": {
										"type": "number",
										"format": "double"
									},
									"author": {
										"type": "string"
									},
									"title": {
										"type": "string"
									}
								},
								"required": [
									"shelfLocation",
									"totalCopies",
									"genre",
									"isbn",
									"author",
									"title"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/libran/addBook/update_book": {
			"put": {
				"operationId": "UpdateBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "API: Update Book\r\nMethod: PUT\r\nURL: /update_book\r\n\r\nBody:\r\n{\r\n  \"book_id\": \"BOOK_ID\",\r\n  \"title\": \"Book Name\",          \r\n  \"author\": \"Author Name\",       \r\n  \"isbn\": \"1234567890\",          \r\n  \"genre\": \"Fiction\",           \r\n  \"totalCopies\": 10,            \r\n  \"availableCopies\": 5,          \r\n  \"shelfLocation\": \"A1\"          \r\n}",
				"tags": [
					"Librarian Add Book"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"shelfLocation": {
										"type": "string"
									},
									"availableCopies": {
										"type": "number",
										"format": "double"
									},
									"totalCopies": {
										"type": "number",
										"format": "double"
									},
									"genre": {
										"type": "string"
									},
									"isbn": {
										"type": "number",
										"format": "double"
									},
									"author": {
										"type": "string"
									},
									"title": {
										"type": "string"
									},
									"book_id": {
										"type": "string"
									}
								},
								"required": [
									"shelfLocation",
									"totalCopies",
									"genre",
									"isbn",
									"author",
									"title",
									"book_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/libran/addBook/get_books": {
			"get": {
				"operationId": "GetBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "API: Get Books\r\nMethod: POST\r\nURL: /get_books\r\n\r\nQuery Params:\r\n{\r\n  \"page\": 1,\r\n  \"limit\": 10,\r\n  \"author\": \"Author Name\",\r\n  \"genre\": \"Genre Name\",\r\n  \"availability\": \"available\" // or \"not_available\"\r\n}",
				"tags": [
					"Librarian Add Book"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "author",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "genre",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "availability",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/libran/addBook/delete_book": {
			"delete": {
				"operationId": "DeleteBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Librarian Add Book"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "book_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}